Upgrade terraform-provider-mongodbatlas to v1.20.0 (#581)
This PR was generated via `$ upgrade-provider
pulumi/pulumi-mongodbatlas`.

---

- Upgrading terraform-provider-mongodbatlas from 1.19.0  to 1.20.0.
	Fixes #580

+++ sdk/dotnet/Outputs/ClusterAdvancedConfiguration.cs
@@ -14,7 +14,7 @@ namespace Pulumi.Mongodbatlas.Outputs
     public sealed class ClusterAdvancedConfiguration
     {
         /// <summary>
-        /// The minimum pre- and post-image retention time in seconds. This option corresponds to the `changeStreamOptions.preAndPostImages.expireAfterSeconds` cluster parameter. Defaults to `-1`(off). This setting controls the retention policy of change stream pre- and post-images. Pre- and post-images are the versions of a document before and after document modification, respectively.`expireAfterSeconds` controls how long MongoDB retains pre- and post-images. When set to -1 (off), MongoDB uses the default retention policy: pre- and post-images are retained until the corresponding change stream events are removed from the oplog. To set the minimum pre- and post-image retention time, specify an integer value greater than zero. Setting this too low could increase the risk of interrupting Realm sync or triggers processing.
+        /// The minimum pre- and post-image retention time in seconds. This option corresponds to the `changeStreamOptions.preAndPostImages.expireAfterSeconds` cluster parameter. Defaults to `-1`(off). This setting controls the retention policy of change stream pre- and post-images. Pre- and post-images are the versions of a document before and after document modification, respectively.`expireAfterSeconds` controls how long MongoDB retains pre- and post-images. When set to -1 (off), MongoDB uses the default retention policy: pre- and post-images are retained until the corresponding change stream events are removed from the oplog. To set the minimum pre- and post-image retention time, specify an integer value greater than zero. Setting this too low could increase the risk of interrupting Realm sync or triggers processing. This parameter is only supported for MongoDB version 6.0 and above.
         /// </summary>
         public readonly int? ChangeStreamOptionsPreAndPostImagesExpireAfterSeconds;
         /// <summary>
@@ -49,7 +49,7 @@ namespace Pulumi.Mongodbatlas.Outputs
         /// Minimum retention window for cluster's oplog expressed in hours. A value of null indicates that the cluster uses the default minimum oplog window that MongoDB Cloud calculates.
         /// * **Note**  A minimum oplog retention is required when seeking to change a cluster's class to Local NVMe SSD. To learn more and for latest guidance see  [`oplogMinRetentionHours`](https://www.mongodb.com/docs/manual/core/replica-set-oplog/#std-label-replica-set-minimum-oplog-size)
         /// </summary>
-        public readonly int? OplogMinRetentionHours;
+        public readonly double? OplogMinRetentionHours;
         /// <summary>
         /// The custom oplog size of the cluster. Without a value that indicates that the cluster uses the default oplog size calculated by Atlas.
         /// </summary>
@@ -83,7 +83,7 @@ namespace Pulumi.Mongodbatlas.Outputs
 
             bool? noTableScan,
 
-            int? oplogMinRetentionHours,
+            double? oplogMinRetentionHours,
 
             int? oplogSizeMb,
