+++ sdk/java/src/main/java/com/pulumi/mongodbatlas/outputs/GetClustersResultAdvancedConfiguration.java
@@ -6,6 +6,7 @@ package com.pulumi.mongodbatlas.outputs;
 import com.pulumi.core.annotations.CustomType;
 import com.pulumi.exceptions.MissingRequiredPropertyException;
 import java.lang.Boolean;
+import java.lang.Double;
 import java.lang.Integer;
 import java.lang.String;
 import java.util.Objects;
@@ -13,7 +14,7 @@ import java.util.Objects;
 @CustomType
 public final class GetClustersResultAdvancedConfiguration {
     /**
-     * @return (Optional) The minimum pre- and post-image retention time in seconds.
+     * @return (Optional) The minimum pre- and post-image retention time in seconds. This parameter is only supported for MongoDB version 6.0 and above. Defaults to `-1`(off).
      * 
      */
     private Integer changeStreamOptionsPreAndPostImagesExpireAfterSeconds;
@@ -59,7 +60,7 @@ public final class GetClustersResultAdvancedConfiguration {
      * @return Minimum retention window for cluster&#39;s oplog expressed in hours. A value of null indicates that the cluster uses the default minimum oplog window that MongoDB Cloud calculates.
      * 
      */
-    private Integer oplogMinRetentionHours;
+    private Double oplogMinRetentionHours;
     /**
      * @return The custom oplog size of the cluster. Without a value that indicates that the cluster uses the default oplog size calculated by Atlas.
      * 
@@ -79,7 +80,7 @@ public final class GetClustersResultAdvancedConfiguration {
 
     private GetClustersResultAdvancedConfiguration() {}
     /**
-     * @return (Optional) The minimum pre- and post-image retention time in seconds.
+     * @return (Optional) The minimum pre- and post-image retention time in seconds. This parameter is only supported for MongoDB version 6.0 and above. Defaults to `-1`(off).
      * 
      */
     public Integer changeStreamOptionsPreAndPostImagesExpireAfterSeconds() {
@@ -139,7 +140,7 @@ public final class GetClustersResultAdvancedConfiguration {
      * @return Minimum retention window for cluster&#39;s oplog expressed in hours. A value of null indicates that the cluster uses the default minimum oplog window that MongoDB Cloud calculates.
      * 
      */
-    public Integer oplogMinRetentionHours() {
+    public Double oplogMinRetentionHours() {
         return this.oplogMinRetentionHours;
     }
     /**
@@ -183,7 +184,7 @@ public final class GetClustersResultAdvancedConfiguration {
         private Boolean javascriptEnabled;
         private String minimumEnabledTlsProtocol;
         private Boolean noTableScan;
-        private Integer oplogMinRetentionHours;
+        private Double oplogMinRetentionHours;
         private Integer oplogSizeMb;
         private Integer sampleRefreshIntervalBiConnector;
         private Integer sampleSizeBiConnector;
@@ -262,7 +263,7 @@ public final class GetClustersResultAdvancedConfiguration {
             return this;
         }
         @CustomType.Setter
-        public Builder oplogMinRetentionHours(Integer oplogMinRetentionHours) {
+        public Builder oplogMinRetentionHours(Double oplogMinRetentionHours) {
             if (oplogMinRetentionHours == null) {
               throw new MissingRequiredPropertyException("GetClustersResultAdvancedConfiguration", "oplogMinRetentionHours");
             }
